ClassicalProbabilisticSystem.jl
This program is a simulation of Classical Probabilistic System. The program asks for the Dynamics matrix (3x3 matrix) and the initial state of the system (1x3 vector).The state is represented by a row vector. It also asks for the number of time clicks.After that, it calculates the state of system before the specified time clicks.