A simulation showing superposition and interference
This is a simulation of a [Quantum] Multi-Slit Experiment \nperformed with Quantum Particles e.g., Photos.

############# ASSUMPTIONS ###############
1) Given the number of slits (n), the number of targets (t) is given by t = 2n + 1.
"##########################################\

The user is asked for the probability amplitudes of the photon from each \nslit to 3 targets behind it. Next, the user enters the time clicks. The program then plots the time evolution of the state vector using Plotly. The program then finds the INTERFERENCE POINTS.